refactor: :truck: componetnsディレクトリを廃止
