+++ packages/app/src/scenes/game/ui/BoardPanel.tsx
@@ -1,11 +1,11 @@
-import { TileButton } from "@/components/buttons/TileButton";
 import { useGame } from "@/game/hooks/useGame";
 import { useUpdateGame } from "@/game/hooks/useUpdateGame";
+import { NeonText } from "@/shared/ui/NeonText";
 import { Panel } from "@/shared/ui/Panel";
 import { cn } from "@/shared/utils/helpers";
 import { type FC, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { NeonText } from "../../shared/ui/NeonText";
+import { TileButton } from "./TileButton";
 
 export const BoardPanel: FC = () => {
   const { board, currentTurn, players } = useGame();
