+++ packages/app/src/scenes/game/ui/GameResetButton.tsx
@@ -1,7 +1,9 @@
 import resetButtonImageUrl from "@/assets/reset.png";
 import { useResetGame } from "@/game/hooks/useResetGame";
 import { useScene } from "@/game/hooks/useScene";
+import { HelpOverlay } from "@/shared/ui/HelpOverlay";
 import { ImageButton } from "@/shared/ui/ImageButton";
+import { NeonText } from "@/shared/ui/NeonText";
 import {
   Tooltip,
   TooltipContent,
@@ -11,8 +13,6 @@ import {
 import { preloadImages } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
-import { HelpOverlay } from "../../shared/ui/HelpOverlay";
-import { NeonText } from "../../shared/ui/NeonText";
 
 preloadImages(resetButtonImageUrl);
