+++ packages/app/src/scenes/mode-select/ui/MultiPlayButton.tsx
@@ -1,26 +1,25 @@
 import { useScene } from "@/game/hooks/useScene";
 import { useUpdatePlayers } from "@/game/hooks/useUpdatePlayers";
 import { analytics } from "@/shared/libs/firebase";
+import { Button } from "@/shared/ui/Button";
 import { logEvent } from "firebase/analytics";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
-import { Button } from "../../shared/ui/Button";
 
-export const SinglePlayButton: FC = () => {
+export const MultiPlayButton: FC = () => {
   const { goToGame } = useScene();
+  const { setMultiPlayer } = useUpdatePlayers();
   const { t } = useTranslation();
-  const { setSinglePlayer } = useUpdatePlayers();
 
   const handleButtonClick = () => {
-    const isPlayerOneCPU = Math.random() < 0.5;
-    setSinglePlayer(isPlayerOneCPU);
+    setMultiPlayer();
     goToGame();
-    logEvent(analytics, "single_play_button_click");
+    logEvent(analytics, "multi_play_button_click");
   };
 
   return (
     <Button variant="panel" onClick={handleButtonClick}>
-      {t("SinglePlayButton.Button")}
+      {t("MultiPlayButton.Button")}
     </Button>
   );
 };
